# src/admin-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/roles": {
      "post": {
        "operationId": "RoleController_create",
        "summary": "Create a new Role",
        "description": "This endpoint is used to create a new role with the provided details. You need to include the role details in the request body, and the API will return the created role object.\n\n**Request Body Parameters**:\n- **role_name** (required): The name of the role to be created.\n\n**How to Use**:\n- Send a POST request to the `/api/v1/admin/roles` endpoint with the role details as a JSON object in the request body.\n- Ensure all required fields are included and correctly formatted.\n\n**Response Details**:\n- On successful creation, the API returns the newly created role object with its generated ID.\n- If any required field is missing or the provided data is invalid, the API responds with a 400 Bad Request error.\n- Server errors or permission issues will result in a 500 Internal Server Error.\n\nThis endpoint allows adding new roles to the system, enabling role management.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "get": {
        "operationId": "RoleController_findAll",
        "summary": "Get a paginated list of Roles",
        "description": "This endpoint retrieves a paginated list of all roles stored in the system. You can filter the results based on various parameters, allowing you to manage and view roles effectively.\n\n**Response Details**:\n- The API returns a JSON object containing a list of roles along with pagination metadata (total count, current page, etc.).\n- If no roles match the criteria, an empty list is returned.\n\nThis endpoint is useful for retrieving and displaying all roles for better role management.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/admin/roles/{id}": {
      "get": {
        "operationId": "RoleController_findOne",
        "summary": "Get Role by ID",
        "description": "This endpoint retrieves the details of a specific role identified by its unique ID. It allows users to view role information in detail.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the role to retrieve.\n\n**Response Details**:\n- On success, the API returns a JSON object with the role’s details.\n- If the role with the specified ID is not found, a 404 Not Found error is returned.\n\nThis endpoint is essential for accessing detailed information about specific roles.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "patch": {
        "operationId": "RoleController_update",
        "summary": "Update Role by ID",
        "description": "This endpoint updates the details of an existing role identified by its unique ID. You need to provide the updated role details in the request body.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the role to update.\n\n**Request Body Parameters**:\n- **role_name** (optional): The updated name of the role.\n\n**Response Details**:\n- On successful update, the API returns the updated role object.\n- If the role is not found, a 404 Not Found error is returned.\n- Validation errors will result in a 400 Bad Request error.\n\nThis endpoint is crucial for maintaining accurate and up-to-date role information.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "delete": {
        "operationId": "RoleController_delete",
        "summary": "Delete Role by ID",
        "description": "This endpoint deletes an existing role from the system using its unique ID. It is a destructive operation that permanently removes the role.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the role to delete.\n\n**Response Details**:\n- On successful deletion, the API returns a confirmation message indicating the role has been deleted.\n- If the role is not found, a 404 Not Found error is returned.\n- Permission errors will result in a 403 Forbidden error.\n\nThis endpoint is essential for managing role lifecycles by allowing deletions of obsolete or incorrect roles.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/admin/roles/{id}/inactivate": {
      "patch": {
        "operationId": "RoleController_inactivate",
        "summary": "Inactivate Role",
        "description": "This endpoint inactivates an existing role identified by its unique ID. It updates the role status to inactive.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the role to inactivate.\n\n**Response Details**:\n- On successful inactivation, the API returns a confirmation message indicating the role has been inactivated.\n- If the role is not found, a 404 Not Found error is returned.\n\nThis endpoint is crucial for managing role statuses effectively.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Admin Service",
    "description": "Admin Service - admin services handle account, project, user and role management",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "role_name": {
            "type": "string",
            "description": "Name of role to be created"
          }
        },
        "required": [
          "role_name"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "role_name": {
            "type": "string",
            "description": "Name of role to be created"
          }
        }
      },
...
    }
  }
}